+++ Neos.Http.Factories/Classes/FlowUploadedFile.php
@@ -1,4 +1,6 @@
 <?php
+declare(strict_types=1);
+
 namespace Neos\Http\Factories;
 
 use GuzzleHttp\Psr7\UploadedFile;
