+++ Neos.Http.Factories/Classes/RequestFactory.php
@@ -1,20 +1,14 @@
 <?php
+declare(strict_types=1);
+
 namespace Neos\Http\Factories;
 
-use GuzzleHttp\Psr7\Request;
 use Psr\Http\Message\RequestFactoryInterface;
-use Psr\Http\Message\RequestInterface;
 
 /**
  *
  */
 class RequestFactory implements RequestFactoryInterface
 {
-    /**
-     * @inheritDoc
-     */
-    public function createRequest(string $method, $uri): RequestInterface
-    {
-        return new Request($method, $uri);
-    }
+    use RequestFactoryTrait;
 }
