+++ Neos.Http.Factories/Classes/RequestFactoryTrait.php
@@ -1,0 +1,22 @@
+<?php
+declare(strict_types=1);
+
+namespace Neos\Http\Factories;
+
+use GuzzleHttp\Psr7\Request;
+use Psr\Http\Message\RequestFactoryInterface;
+use Psr\Http\Message\RequestInterface;
+
+/**
+ *
+ */
+trait RequestFactoryTrait
+{
+    /**
+     * @inheritDoc
+     */
+    public function createRequest(string $method, $uri): RequestInterface
+    {
+        return new Request($method, $uri);
+    }
+}
