+++ Neos.Http.Factories/Classes/ResponseFactory.php
@@ -1,20 +1,14 @@
 <?php
+declare(strict_types=1);
+
 namespace Neos\Http\Factories;
 
-use GuzzleHttp\Psr7\Response;
 use Psr\Http\Message\ResponseFactoryInterface;
-use Psr\Http\Message\ResponseInterface;
 
 /**
  *
  */
 class ResponseFactory implements ResponseFactoryInterface
 {
-    /**
-     * @inheritDoc
-     */
-    public function createResponse(int $code = 200, string $reasonPhrase = ''): ResponseInterface
-    {
-        return new Response($code, [], null, '1.1', $reasonPhrase);
-    }
+    use ResponseFactoryTrait;
 }
