+++ Neos.Http.Factories/Classes/ResponseFactoryTrait.php
@@ -1,0 +1,22 @@
+<?php
+declare(strict_types=1);
+
+namespace Neos\Http\Factories;
+
+use GuzzleHttp\Psr7\Response;
+use Psr\Http\Message\ResponseFactoryInterface;
+use Psr\Http\Message\ResponseInterface;
+
+/**
+ *
+ */
+trait ResponseFactoryTrait
+{
+    /**
+     * @inheritDoc
+     */
+    public function createResponse(int $code = 200, string $reasonPhrase = ''): ResponseInterface
+    {
+        return new Response($code, [], null, '1.1', $reasonPhrase);
+    }
+}
