+++ Neos.Http.Factories/Classes/UriFactory.php
@@ -1,20 +1,14 @@
 <?php
+declare(strict_types=1);
+
 namespace Neos\Http\Factories;
 
-use GuzzleHttp\Psr7\Uri;
 use Psr\Http\Message\UriFactoryInterface;
-use Psr\Http\Message\UriInterface;
 
 /**
  *
  */
 class UriFactory implements UriFactoryInterface
 {
-    /**
-     * @inheritDoc
-     */
-    public function createUri(string $uri = ''): UriInterface
-    {
-        return new Uri($uri);
-    }
+    use UriFactoryTrait;
 }
