+++ Neos.Http.Factories/Classes/UriFactoryTrait.php
@@ -1,0 +1,22 @@
+<?php
+declare(strict_types=1);
+
+namespace Neos\Http\Factories;
+
+use GuzzleHttp\Psr7\Uri;
+use Psr\Http\Message\UriFactoryInterface;
+use Psr\Http\Message\UriInterface;
+
+/**
+ *
+ */
+trait UriFactoryTrait
+{
+    /**
+     * @inheritDoc
+     */
+    public function createUri(string $uri = ''): UriInterface
+    {
+        return new Uri($uri);
+    }
+}
